Adding learning to rank algorithm
