+++ src/location/location_recommender.py
@@ -41,8 +41,14 @@ class LocationRecommender:
         for event_id in potential_events:
             lat = events_info[event_id]["lat"]
             lon = events_info[event_id]["lon"]
+            # similarity_scores.append(np.exp(kde.score([np.array([lat, lon]).T])))
             similarity_scores.append(simscores[member_id][event_id])
             simscores[member_id][event_id] = np.exp(kde.score([np.array([lat, lon]).T]))
+
+        # similarity_scores = np.array(similarity_scores)
+        # top_indices = similarity_scores.argsort()[:-6:-1]
+        # for i in top_indices:
+        #     simscores[member_id][potential_events[i]] = similarity_scores[i]
         #print simscores
         #TEST: Pick top 5 similar scores. Print all events. Print top 5 events.
         #similarity_scores = np.array(similarity_scores)
