+++ src/temporal/time_recommender.py
@@ -51,6 +51,10 @@ class TimeRecommender:
         for i in xrange(len(potential_events)):
             simscores[member_id][potential_events[i]] = similarity_scores[i]
 
+        # top_indices = similarity_scores.argsort()[:-6:-1]
+        # for i in top_indices:
+        #     simscores[member_id][potential_events[i]] = similarity_scores[i]
+
         # top_indices = similarity_scores.argsort()[:-5:-1]
         # top_5_recommended_events = []
         # for i in top_indices:
